+++ src/errors/BaseError.ts
