+++ src/errors/InvalidServerFeaturesError.ts
