+++ src/errors/RecordLockedError.ts
