+++ src/errors/__tests__/ConnectionManagerError.test.ts
