+++ src/errors/__tests__/DbServerError.test.ts
@@ -1,4 +1,4 @@
-import { dbErrors } from '../../../constants';
+import { dbErrors } from '../../constants';
 import DbServerError from '../DbServerError';
 
 // use one of the codes from the errors table
