+++ src/errors/__tests__/InvalidServerFeaturesError.test.ts
