+++ src/errors/__tests__/RecordLockedError.test.ts
