+++ src/errors/__tests__/RecordVersionError.test.ts
