+++ src/index.ts
@@ -1,4 +1,4 @@
 export { default as Connection } from './Connection';
 export { default as Document } from './Document';
 export { default as Schema } from './Schema';
-export * as Errors from './shared/errors';
+export * as Errors from './errors';
