+++ src/shared/schemaType/ISOCalendarDateType.ts
@@ -1,6 +1,6 @@
 import moment from 'moment';
 import { ISOCalendarDateFormat, mvEpoch } from '../../constants';
-import { TransformDataError } from '../errors';
+import { TransformDataError } from '../../errors';
 import { handleTypeValidation } from '../utils';
 import BaseDateType from './BaseDateType';
 import type { ScalarTypeConstructorOptions } from './BaseScalarType';
