+++ src/shared/schemaType/__tests__/ISOCalendarDateTimeType.test.ts
@@ -1,4 +1,4 @@
-import { TransformDataError } from '../../errors';
+import { TransformDataError } from '../../../errors';
 import type { SchemaTypeDefinitionISOCalendarDateTime } from '../ISOCalendarDateTimeType';
 import ISOCalendarDateTimeType from '../ISOCalendarDateTimeType';
