+++ src/shared/schemaType/__tests__/ISOCalendarDateType.test.ts
@@ -1,4 +1,4 @@
-import { TransformDataError } from '../../errors';
+import { TransformDataError } from '../../../errors';
 import type { SchemaTypeDefinitionISOCalendarDate } from '../ISOCalendarDateType';
 import ISOCalendarDateType from '../ISOCalendarDateType';
